br/lightning: split restore_test.go into multiple files
close #32036
